Export more types on API surface

+++ leap-server/src/provision/storage.rs
@@ -31,7 +31,7 @@ pub struct BlockDevice {
     #[serde(rename = "type")]
     pub ty: BlockDeviceType,
 
-    pub size: usize,
+    pub size: u64,
 
     pub mountpoints: Vec<String>,
